Added keepout, drill, and layer options
Split svg2mod/svg2mod into four files: cli, exporter, importer, svg2mod
