+++ .linter/cleanup.rc
@@ -75,7 +75,6 @@ enable=c-extension-no-member,
 		no-else-continue,
 		no-else-raise,
 		useless-object-inheritance,
-		too-few-public-methods,
 		too-many-arguments,
 		too-many-branches,
 		too-many-instance-attributes,
