+++ .linter/custom_dict
@@ -44,3 +44,4 @@ Regex
 thru
 Faux
 getLogger
+keepout
