+++ setup.py
@@ -48,7 +48,7 @@ setup(
     author_email='',
     url='https://github.com/svg2mod/svg2mod',
     packages=setuptools.find_packages(),
-    entry_points={'console_scripts':['svg2mod = svg2mod.svg2mod:main']},
+    entry_points={'console_scripts':['svg2mod = svg2mod.cli:main']},
     package_dir={'svg2mod':'svg2mod'},
     include_package_data=True,
     package_data={'kipart': ['*.gif', '*.png']},
@@ -58,8 +58,10 @@ setup(
     zip_safe=False,
     keywords='svg2mod, KiCAD, inkscape',
     classifiers=[
-        'Development Status :: 3 - Alpha',
+        'Development Status :: 4 - Beta',
         'Intended Audience :: Science/Research',
+        'Intended Audience :: Developers',
+        'Intended Audience :: Manufacturing',
         'License :: OSI Approved :: GNU General Public License v2 or later (GPLv2+)',
         'Natural Language :: English',
         'Programming Language :: Python :: 3',
