+++ svg2mod/__init__.py
@@ -1,0 +1,23 @@
+# Copyright (C) 2022 -- svg2mod developers < GitHub.com / svg2mod >
+
+# This program is free software; you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation; either version 2 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License along
+# with this program; if not, write to the Free Software Foundation, Inc.,
+# 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
+
+'''
+This module contains the necessary tools to convert from
+the svg objects provided from the svg2mod.svg module to
+KiCad file formats.
+This currently supports both the pretty format and
+the legacy mod format.
+'''
