+++ svg2mod/importer.py
@@ -1,0 +1,71 @@
+# Copyright (C) 2021 -- svg2mod developers < GitHub.com / svg2mod >
+
+# This program is free software; you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation; either version 2 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License along
+# with this program; if not, write to the Free Software Foundation, Inc.,
+# 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
+
+'''
+Svg2ModImport is responsible for basic parsing
+of svg layers to be used with an instance of
+Svg2ModExport.
+'''
+
+import logging
+
+from svg2mod import svg
+
+
+class Svg2ModImport:
+    ''' An importer class to read in target svg,
+    parse it, and keep only layers on interest.
+    '''
+
+
+    def _prune_hidden( self, items = None ):
+
+        if items is None:
+
+            items = self.svg.items
+            self.svg.items = []
+
+        for item in items:
+
+            if not isinstance( item, svg.Group ):
+                continue
+
+            if item.hidden :
+                logging.warning("Ignoring hidden SVG layer: {}".format( item.name ) )
+            elif item.name != "":
+                self.svg.items.append( item )
+
+            if item.items:
+                self._prune_hidden( item.items )
+
+    def __init__( self, file_name=None, module_name="svg2mod", module_value="G***", ignore_hidden_layers=False ):
+
+        self.file_name = file_name
+        self.module_name = module_name
+        self.module_value = module_value
+
+        if file_name:
+            logging.getLogger("unfiltered").info( "Parsing SVG..." )
+
+            self.svg = svg.parse( file_name)
+            logging.info("Document scaling: {} units per pixel".format(self.svg.viewport_scale))
+        if ignore_hidden_layers:
+            self._prune_hidden()
+
+
+    #------------------------------------------------------------------------
+
+#----------------------------------------------------------------------------
